# public/wapd-api-doc.json
{
  "swagger": "2.0",
  "info": {
    "description": "Wapd api document for front-end developer",
    "version": "0.0.1",
    "title": "Wapd",
    "contact": {
      "email": "uu9924079@gmail.com"
    }
  },
  "host": "localhost:3000",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
...
  "paths": {
...
    "/api/orders/{id}/checkout": {
      "get": {
        "tags": ["orders"],
        "summary": "Display an order's checkout",
        "description": "",
        "operationId": "orderController.getCheckout",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of order to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object",
            "schema": {
              "type": "object",
              "properties": {
                "orderItems": {
                  "type": "array",
                  "example": "[{orderItem Object}, {orderItem Object}]"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/orders/checkout": {
      "post": {
        "tags": ["admins"],
        "summary": "Post order and get to payment page",
        "description": "",
        "operationId": "orderController.postCheckout",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Post order and get to payment page",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "receiverName": {
                  "type": "string"
                },
                "receiverPhone": {
                  "type": "string"
                },
                "receiverAddress": {
                  "type": "string"
                },
                "receiverEmail": {
                  "type": "string"
                },
                "total": {
                  "type": "integer"
                },
                "orderId": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/orders/{id}/payment": {
      "get": {
        "tags": ["orders"],
        "summary": "Display payment of a order",
        "description": "",
        "operationId": "orderController.getPayment",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of order to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/paymentcomplete": {
      "get": {
        "tags": ["orders"],
        "summary": "Display payment complete page",
        "description": "",
        "operationId": "orderController.getPaymentComplete",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Status",
            "in": "query",
            "description": "status",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "orderId",
            "in": "query",
            "description": "orderId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/admins/orders": {
      "get": {
        "tags": ["admins"],
        "summary": "Display all orders",
        "description": "",
        "operationId": "adminController.getOrders",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ],
        "responses": {
          "200": {
            "description": "return all orders",
            "schema": {
              "type": "object",
              "properties": {
                "orders": {
                  "type": "array",
                  "example": "[{order object}, {order object}]"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/admins/orders/transition": {
      "post": {
        "tags": ["orders"],
        "summary": "Check a transition",
        "description": "",
        "operationId": "orderController.postTransition",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "send sn, amt to check an existing order",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "sn": {
                  "type": "integer"
                },
                "amt": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}